+++ pkg/networks/networks.go
@@ -20,3 +20,12 @@ func NewClient(ctx context.Context, oauth2 *clientcredentials.Config, overrideSv
 
 	return clients.NewClient[Network](ctx, svr, "/v1/networks", oauth2, scopeNetworksModify)
 }
+
+func NewDisplayClient(ctx context.Context, oauth2 *clientcredentials.Config, overrideSvr ...string) (*clients.ChildClient[NetworkDisplay], error) {
+	svr := serverUrl
+	if len(overrideSvr) > 0 && overrideSvr[0] != "" {
+		svr = overrideSvr[0]
+	}
+
+	return clients.NewChildClient[NetworkDisplay](ctx, svr, "/v1/networks", "/displays", oauth2, scopeNetworksModify)
+}